sigma min max limit 구현 완료

+++ include/caffe/loss_layers.hpp
@@ -850,8 +850,8 @@ protected:
 	
 	int class_size;
 	int data_dim;
-	int sigma_min;
-	int sigma_max;
+	float sigma_min;
+	float sigma_max;
 	bool visualize_;
 
 	Blob<Dtype> diff_;					//(mu-t)
